Moved the refactored methods into the GameEngine and removed the Interface
